Feat: Add `useHasNoDecisionsMethod` hook
