+++ src/components/v5/common/ActionSidebar/hooks/index.ts
@@ -8,5 +8,4 @@ export * from './useGetColonyAction.ts';
 export * from './useRemoveTxParamOnClose.ts';
 export * from './useSidebarActionForm.ts';
 export * from './useReputationValidation.ts';
-export * from './useDecisionMethods.ts';
-export * from './useHasActionPermissions/index.ts';
+export * from './permissions/index.ts';
