+++ src/components/v5/common/ActionSidebar/hooks/permissions/index.ts
@@ -1,1 +1,2 @@
 export * from './useHasActionPermissions.ts';
+export * from './useHasNoDecisionMethods.ts';
